+++ examples/evtest.rs
@@ -5,11 +5,11 @@ mod _pick_device;
 
 fn main() {
     let mut d = _pick_device::pick_device();
-    println!("{}", d);
+    println!("{d}");
     println!("Events:");
     loop {
         for ev in d.fetch_events().unwrap() {
-            println!("{:?}", ev);
+            println!("{ev:?}");
         }
     }
 }