+++ examples/evtest_nonblocking.rs
@@ -18,19 +18,18 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
         fcntl::{FcntlArg, OFlag},
         sys::epoll,
     };
-    use std::os::unix::io::AsRawFd;
+    use std::os::fd::{AsRawFd, FromRawFd, OwnedFd};
 
     let mut d = _pick_device::pick_device();
-    println!("{}", d);
+    println!("{d}");
 
     let raw_fd = d.as_raw_fd();
     // Set nonblocking
     nix::fcntl::fcntl(raw_fd, FcntlArg::F_SETFL(OFlag::O_NONBLOCK))?;
 
     // Create epoll handle and attach raw_fd
-    let epoll_fd = crate::epoll::Epoll::new(epoll::epoll_create1(
-        epoll::EpollCreateFlags::EPOLL_CLOEXEC,
-    )?);
+    let epoll_fd = epoll::epoll_create1(epoll::EpollCreateFlags::EPOLL_CLOEXEC)?;
+    let epoll_fd = unsafe { OwnedFd::from_raw_fd(epoll_fd) };
     let mut event = epoll::EpollEvent::new(epoll::EpollFlags::EPOLLIN, 0);
     epoll::epoll_ctl(
         epoll_fd.as_raw_fd(),
@@ -47,7 +46,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
         match d.fetch_events() {
             Ok(iterator) => {
                 for ev in iterator {
-                    println!("{:?}", ev);
+                    println!("{ev:?}");
                 }
             }
             Err(e) if e.kind() == std::io::ErrorKind::WouldBlock => {
@@ -55,37 +54,10 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
                 epoll::epoll_wait(epoll_fd.as_raw_fd(), &mut events, -1)?;
             }
             Err(e) => {
-                eprintln!("{}", e);
+                eprintln!("{e}");
                 break;
             }
         }
     }
     Ok(())
-}
-
-#[cfg(target_os = "linux")]
-mod epoll {
-    use std::os::unix::io::{AsRawFd, RawFd};
-
-    // The rest here is to ensure the epoll handle is cleaned up properly.
-    // You can also use the epoll crate, if you prefer.
-    pub(crate) struct Epoll(RawFd);
-
-    impl Epoll {
-        pub(crate) fn new(fd: RawFd) -> Self {
-            Epoll(fd)
-        }
-    }
-
-    impl AsRawFd for Epoll {
-        fn as_raw_fd(&self) -> RawFd {
-            self.0
-        }
-    }
-
-    impl Drop for Epoll {
-        fn drop(&mut self) {
-            let _ = nix::unistd::close(self.0);
-        }
-    }
 }